+++ membership/options.lua
@@ -1,10 +1,13 @@
 local options = {}
+local log = require('log')
+local cbc = require('crypto').cipher.aes256.cbc
 
-options.STATUS_NAMES = {'alive', 'suspect', 'dead', 'left'}
+options.STATUS_NAMES = {'alive', 'suspect', 'dead', 'non-decryptable', 'left'}
 options.ALIVE = 1
 options.SUSPECT = 2
 options.DEAD = 3
-options.LEFT = 4
+options.NONDECRYPTABLE = 4
+options.LEFT = 5
 
 options.PROTOCOL_PERIOD_SECONDS = 1.0 -- denoted as T' in SWIM paper
 options.ACK_TIMEOUT_SECONDS = 0.200 -- ack timeout
@@ -21,15 +24,45 @@ function options.set_advertise_uri(uri)
     rawset(options, 'advertise_uri', uri)
 end
 
+function options.get_encryption_key()
+    return options.encryption_key
+end
+
 function options.set_encryption_key(key)
     if key == nil then
         rawset(options, 'encryption_key', nil)
+        log.info('Membership encryption disabled')
     else
         if key:len() < 32 then
             rawset(options, 'encryption_key', key:rjust(32))
-        elseif key:len() > 32 then
+        else
             rawset(options, 'encryption_key', key:sub(1, 32))
         end
+        log.info('Membership encryption enabled')
+    end
+end
+
+function options.encrypt(msg)
+    if not options.encryption_key then
+        return msg, nil
+    else
+        return cbc.encrypt(
+            msg,
+            options.encryption_key,
+            options.ENCRYPTION_INIT
+        )
+    end
+end
+
+function options.decrypt(msg)
+    if not options.encryption_key then
+        return msg, nil
+    else
+        return cbc.decrypt(
+            msg,
+            options.encryption_key,
+            options.ENCRYPTION_INIT
+        )
     end
 end